fix: created TraefikMeDomainStandaloneService for job scheduling of updateTraefikMeCertificate
