+++ src/server.ts
@@ -8,7 +8,7 @@ import dataAccess from './server/adapter/db.client'
 import { FancyConsoleUtils } from './shared/utils/fancy-console.utils'
 import { Constants } from './shared/utils/constants'
 import backupService from './server/services/standalone-services/backup.service'
-import traefikMeDomainService from './server/services/standalone-services/traefik-me-domain.service'
+import traefikMeDomainStandaloneService from './server/services/standalone-services/traefik-me-domain-standalone.service'
 
 // Source: https://nextjs.org/docs/app/building-your-application/configuring/custom-server
 
@@ -53,7 +53,7 @@ async function initializeNextJs() {
     }
 
     await backupService.registerAllBackups();
-    traefikMeDomainService.configureSchedulingForTraefikMeCertificateUpdate();
+    traefikMeDomainStandaloneService.configureSchedulingForTraefikMeCertificateUpdate();
 
     const app = next({ dev });
     const handle = app.getRequestHandler();
