+++ src/server/services/project.service.ts
@@ -6,7 +6,7 @@ import { KubeObjectNameUtils } from "../utils/kube-object-name.utils";
 import deploymentService from "./deployment.service";
 import namespaceService from "./namespace.service";
 import buildService from "./build.service";
-import traefikMeDomainService from "./standalone-services/traefik-me-domain.service";
+import traefikMeDomainStandaloneService from "./standalone-services/traefik-me-domain-standalone.service";
 
 class ProjectService {
 
@@ -70,7 +70,7 @@ class ProjectService {
         } finally {
             revalidateTag(Tags.projects());
         }
-        await traefikMeDomainService.updateTraefikMeCertificate();
+        await traefikMeDomainStandaloneService.updateTraefikMeCertificate();
         return savedItem;
     }
 }
